# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SNT Portal API
  version: 1.0.0
  description: |
    Unified API for resident/chairman portal.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/news:
    get:
      summary: List news
      security:
        - bearerAuth: []
    post:
      summary: Create draft news (chairman)
      security:
        - bearerAuth: []
  /api/v1/documents:
    get:
      summary: List documents
      security:
        - bearerAuth: []
    post:
      summary: Upload document metadata (chairman)
      security:
        - bearerAuth: []
  /api/v1/forum/threads:
    get:
      summary: List threads
      security:
        - bearerAuth: []
    post:
      summary: Create thread
      security:
        - bearerAuth: []
...
  /api/v1/map/layers:
    get:
      summary: List map layers with objects
      security:
        - bearerAuth: []
    post:
      summary: Create map layer (chairman)
      security:
        - bearerAuth: []
  /api/v1/incidents:
    get:
      summary: List incidents
      security:
        - bearerAuth: []
    post:
      summary: Create incident
      security:
        - bearerAuth: []
  /api/v1/meetings:
    get:
      summary: List meetings
      security:
        - bearerAuth: []
    post:
      summary: Create meeting (chairman)
      security:
        - bearerAuth: []
...